Use EH iso codes from NaturalEarth
Natural Earth does not provide ISO codes for some large countries, in
particular France and Norway. This is a conscious decision that is
rooted in an imprecise or overlapping definition of country codes by
ISO, where Natural Earth follows a stricter mode that is more fine
grained.
For users who are ok with an imprecise definition and who would need
the ISO code for Paris's geo coordinates to be "FR" rather than "-99",
they provide the ISO_EH fields (fuzzy fields).

This commit addresses this issue by switching to the "EH" fields and
recreates the dataset, replacing some of the "-99" values with the
commonly accepted ISO codes.

Context for this issue is upstream at https://github.com/nvkelso/natural-earth-vector/issues/268

This has been discussed for this codebase before, and is currently in
regression https://github.com/datasets/geo-countries/pull/6